Improved Build Manager Raycasting
Layermasks! Wow, that helps! Now I check a "Building" layer for a hit,
and check the "Turrets" layer for no hits... so it's hard to build on
top of other turrets now. It has room for improvement, because you can
still overlap by raycasting just right. Perhaps I need to look up a way
to bounding box check using the raycast as a center point next.